Add tests for the moving hand cart and farmerNPC's implementation
The current tests do not include the moving handcarts implementation.
Also doesn't include farmerNPC's behaviour when handcart is summoned.

These new set of tests allow the handcart to be tested:
To see whether it can be appropriatley pushed.
To check what the FarmerNPC says to the player.
To check if the player is a high enough level to summon the handcart.

Tests that the handcart can be pushed in all directions.
Tests the coordinates of the "push" makes the handcart move correctly.
Tests if the handcart reacts appropriately to being open or closed.
The player should not be able to push the hand cart if it is open.
Tests the FarmerNPCs conversation with the player.

To make these tests pass, implement:
HandCart, FarmerNPC and potentially StendhalRPZone.