Adds Third test for Disguises.
This test checks that the player is made visible when player is wearing
the correct armour piece but leaving Nalwor zone. The test picks a
Nalwor zone at random and generates a new player and equips the armour
piece. Then we teleport the player to any other zone than Nalwor and
check that the player is visible.