Replaced second test for disguises.
This checks that the player is made invisible if wearing the correct
piece of armour and entering in a Nalwor zone from another zone. We pick
a zone other than Nalwor and the test then equips the armour piece.
After this the test teleports the player to a Nalwor zone and checks
that the player is invisible.